% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_env.R
\name{activate_mutater_env}
\alias{activate_mutater_env}
\title{Activate the mutateR Python environment}
\usage{
activate_mutater_env(envname = "r-mutater")
}
\arguments{
\item{envname}{Character. Name of the environment (default "r-mutater").}
}
\value{
Logical TRUE if successful, FALSE otherwise.
}
\description{
Points the current R session to the mutateR Conda environment.
Automatically handles RETICULATE_PYTHON conflicts.
}
