% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_exon_phase.R
\name{check_exon_phase}
\alias{check_exon_phase}
\title{Check exon pair phase compatibility}
\usage{
check_exon_phase(exon_info, include_contiguous = FALSE)
}
\arguments{
\item{exon_info}{data.frame or GRanges with columns \code{rank}, \code{start_phase}, \code{end_phase}.
Output from get_exon_structures() can be adapted by renaming "phase" to "start_phase".}

\item{include_contiguous}{Logical. Whether to include contiguous exon pairs
(default FALSE; only non-contiguous pairs are shown).}
}
\value{
A data.frame of compatible exon pairs with columns:
\itemize{
\item \code{exon_5p} (the upstream exon rank)
\item \code{exon_3p} (the downstream exon rank)
\item \code{end_phase_5p}
\item \code{start_phase_3p}
\item \code{compatible} (TRUE/FALSE)
}
}
\description{
Determines which pairs of exons are phase-compatible, meaning that
if the exons between them were deleted, the original reading frame would be preserved.
}
\examples{
\dontrun{
exons <- data.frame(rank=1:5,
                    start_phase=c(0,2,0,1,1),
                    end_phase=c(2,0,1,1,-1))
check_exon_phase(exons)
}
}
