% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_valid_grnas.R
\name{filter_valid_grnas}
\alias{filter_valid_grnas}
\title{Integrate gRNA scoring data with exon‑pair compatibility}
\usage{
filter_valid_grnas(
  exon_gr,
  genome,
  species,
  score_method = c("ruleset1", "azimuth", "deephf", "deepcpf1", "deepspcas9", "ruleset3"),
  scored_grnas = NULL,
  tracr = "Chen2013"
)
}
\arguments{
\item{exon_gr}{GRanges from get_exon_structures(output="GRanges").}

\item{genome}{BSgenome object for the relevant species.}

\item{species}{Character, e.g. "hsapiens".}

\item{score_method}{Character. One of "ruleset1", "azimuth", "deepspcas9", "deephf", "ruleset3 (Cas9) or "deepcpf1" (Cas12a).}

\item{scored_grnas}{Precomputed on-target scores passed from elsewhere (default NULL, evokes scoring).}

\item{tracr}{Character, either "Chen2013" (default) or "Hsu2013" - only for ruleset3 scoring.}
}
\value{
GRanges of scored, biologically admissible gRNAs.
}
\description{
Returns scored Cas9 gRNAs that lie within exons
participating in phase‑compatible exon pairs or predicted
terminal‑exon frameshift cases (permitted PTC deletions).
}
