% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cas12a_sites.R
\name{find_cas12a_sites}
\alias{find_cas12a_sites}
\title{Find Cas12a (Cpf1) target sites in exons}
\usage{
find_cas12a_sites(exon_gr, genome, pam = "TTTV", protospacer_length = 23)
}
\arguments{
\item{exon_gr}{GRanges from get_exon_structures(..., output = "GRanges").}

\item{genome}{BSgenome object (e.g., BSgenome.Hsapiens.UCSC.hg38).}

\item{pam}{PAM (default "TTTV"). IUPAC ambiguity codes are supported.}

\item{protospacer_length}{Protospacer length downstream of PAM (default 23 bp).}
}
\value{
GRanges of Cas12a target sites with metadata columns:
exon_rank, protospacer_sequence, pam_sequence, target_sequence,
sequence_context (34 bp), cut_site.
}
\description{
Scans exon sequences (both strands) for Cas12a gRNA sites (default PAM "TTTV")
Retrieves full genomic context needed for DeepCpf1 scoring (34 bp) even if sites are near exon boundaries.
}
