% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cas9_sites.R
\name{find_cas9_sites}
\alias{find_cas9_sites}
\title{Find Cas9 target sites in exons}
\usage{
find_cas9_sites(exon_gr, genome, pam = "NGG", protospacer_length = 20)
}
\arguments{
\item{exon_gr}{GRanges from get_exon_structures(..., output="GRanges").
Must contain metadata column \code{rank}.}

\item{genome}{BSgenome object (e.g. BSgenome.Hsapiens.UCSC.hg38).}

\item{pam}{PAM (default "NGG"). IUPAC ambiguity codes are supported.}

\item{protospacer_length}{Protospacer length upstream of PAM (default 20 bp for Cas9).}
}
\value{
GRanges object of candidate scorable Cas9 binding sites with relevant metadata.
}
\description{
Scans exon sequences on both strands for NGG PAMs.
Excludes gRNA sequences where the flanking -4/+3 nucleotides would run out of an exon boundary (for on-target scoring purposes).
Outputs both + and - strand hits, each with a correctly oriented 30‑nt
sequence_context = [4-nt flank][20-nt protospacer][3-nt PAM][3-nt flank].
}
