% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_info.R
\name{get_gene_info}
\alias{get_gene_info}
\title{Retrieve transcript information for a selected gene from a given species}
\usage{
get_gene_info(gene_id, species, id_type = c("symbol", "ensembl_gene_id"))
}
\arguments{
\item{gene_id}{Character. Identifier for selected gene of interest: either gene symbol
(via external_gene_name) or an Ensembl gene ID (e.g., ENSG...).}

\item{species}{Character. Species short name (e.g., "hsapiens", "drerio", "mmusculus", etc.).}

\item{id_type}{Character. One of "symbol" (default) or "ensembl_gene_id".}
}
\value{
A list with:
\itemize{
\item canonical: data.frame of canonical transcript(s), or NULL if no transcript annotated as canonical for the selected gene
\item all: full data.frame of transcripts, canonical (if present) raised to top
}
}
\description{
Queries Ensembl via biomaRt to retrieve transcript annotation for a given
gene symbol or Ensembl gene ID in a specified species. Canonical
transcript(s), if any, are flagged and the output separated into a list of
two data.frames (all transcripts, and just the canonical transcript
(if applicable)).
}
\examples{
\dontrun{
tx_info <- get_gene_info("TP53", "hsapiens")
tx_info$canonical    # canonical transcript(s)
tx_info$all          # all transcripts
}
}
