% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutateR_viewer.R
\name{mutateR_viewer}
\alias{mutateR_viewer}
\title{Interactive viewer for mutateR gRNA design heatmaps}
\usage{
mutateR_viewer(plot_obj)
}
\arguments{
\item{plot_obj}{Plotly object returned in the mutateR output list
(element \code{$plot}); must have attribute \code{"pairs_data"}.}
}
\description{
Launches a small Shiny gadget showing the interactive Plotly heatmap.
Clicking a cell filters the gRNA table below.
}
