% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grna_heatmap.R
\name{plot_grna_heatmap}
\alias{plot_grna_heatmap}
\title{Plot exon‑phase compatibility as a discrete heatmap
(helper function for large-gene visualisation)}
\usage{
plot_grna_heatmap(exon_gr, transcript_id = "Transcript", gene_symbol = NULL)
}
\arguments{
\item{exon_gr}{GRanges from get_exon_structures(output="GRanges")}

\item{transcript_id}{Character. Transcript label.}

\item{gene_symbol}{Character. Gene label.}
}
\value{
ggplot object
}
\description{
Plot exon‑phase compatibility as a discrete heatmap
(helper function for large-gene visualisation)
}
