% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_grnas.R
\name{score_grnas}
\alias{score_grnas}
\title{Score individual gRNAs using crisprScore models}
\usage{
score_grnas(
  grna_gr,
  method = c("ruleset1", "azimuth", "ruleset3", "deepspcas9", "deephf", "deepcpf1",
    "enpamgb")
)
}
\arguments{
\item{grna_gr}{GRanges returned by find_cas9_sites() or find_cas12a_sites().}

\item{method}{Scoring model name.  Default "ruleset1".}
}
\value{
GRanges with added metadata column \code{ontarget_score} (numeric)
}
\description{
Handles RuleSet1 (Doench et al., Nat Biotech (2014)) and Azimuth (Doench et al. Nat Biotech (2016) -- in progress) with proper PAM alignment.
For models unsupported on Windows, returns NA scores and a warning.
}
