% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_enpamgb.R
\name{validate_enpamgb}
\alias{validate_enpamgb}
\title{Validate enPAM+GB installation}
\usage{
validate_enpamgb()
}
\value{
Logical; TRUE if enPAM+GB is working correctly, FALSE otherwise.
}
\description{
Tests that the enPAM+GB model and dependencies are correctly installed.
}
