% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_rs3.R
\name{validate_rs3}
\alias{validate_rs3}
\title{Validate RS3 installation}
\usage{
validate_rs3()
}
\value{
Logical; TRUE if RS3 is working correctly, FALSE otherwise.
}
\description{
Tests that the RS3 Python module is correctly installed and functional.
}
