% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_frameshift_ptc.R
\name{check_frameshift_ptc}
\alias{check_frameshift_ptc}
\title{Assess if an deletion of exon(s) would cause a frameshift/PTC}
\usage{
check_frameshift_ptc(exon_info, exon_5p, exon_3p)
}
\arguments{
\item{exon_info}{data.frame with \code{rank} and \code{exon_cds_length}}

\item{exon_5p}{Integer. Upstream exon rank.}

\item{exon_3p}{Integer. Downstream exon rank.}
}
\value{
List with frameshift (logical), ptc_flag (logical),
terminal_exon_case (logical), and deleted_length (integer).
}
\description{
Assess if an deletion of exon(s) would cause a frameshift/PTC
}
