% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_valid_grnas.R
\name{filter_valid_grnas}
\alias{filter_valid_grnas}
\title{Integrate gRNA scoring data with exon‑pair compatibility}
\usage{
filter_valid_grnas(
  exon_gr,
  genome,
  species,
  nuclease = c("Cas9", "Cas12a", "enCas12a"),
  score_method = c("ruleset1", "azimuth", "deepspcas9", "deephf", "ruleset3", "deepcpf1",
    "enpamgb"),
  scored_grnas = NULL,
  tracr = "Chen2013",
  deephf_var = c("wt_u6", "wt_t7", "wt", "esp", "hf")
)
}
\arguments{
\item{exon_gr}{GRanges from get_exon_structures(output="GRanges").}

\item{genome}{BSgenome object for the relevant species.}

\item{species}{Character, e.g. "hsapiens".}

\item{nuclease}{Character. One of "Cas9", "Cas12a" or "enCas12a".}

\item{score_method}{Character. One of "ruleset1", "azimuth", "deepspcas9",
"deephf", "ruleset3" (Cas9), "deepcpf1" (Cas12a), or "enpamgb" (enCas12a).}

\item{scored_grnas}{Precomputed on-target scores passed from elsewhere (default NULL, evokes scoring).}

\item{tracr}{Character, either "Chen2013" (default) or "Hsu2013" - only for ruleset3 scoring.}

\item{deephf_var}{Character. For DeepHF scoring; one of "wt", "wt_u6" (default), "wt_t7", "esp", or "hf".
See \code{\link{recommend_deephf_model}} for guidance on model selection.}
}
\value{
GRanges of scored, mutateR-allowed gRNAs.
}
\description{
Returns scored Cas9/Cas12a gRNAs that lie within exons
participating in phase‑compatible exon pairs or predicted
terminal‑exon PTC cases.
}
